// apps/mobile/src/components/EarningsChart.d.ts
import React from 'react';
import { ChartData } from '../types';
interface EarningsChartProps {
  data: ChartData;
  type?: 'line' | 'bar';
  title?: string;
  subtitle?: string;
  height?: number;
}
export declare const EarningsChart: React.FC<EarningsChartProps>;
export {};
//# sourceMappingURL=EarningsChart.d.ts.map